chore: Move `run_parallel_tasks` to the `service` crate so it can be reused in other routes
It does not depend on any HTTP REST concept or type so it has its place in `service::util`.
