+++ crates/service/Cargo.toml
@@ -14,6 +14,7 @@ chrono = { workspace = true }
 email_address = { workspace = true }
 figment = { workspace = true }
 form_urlencoded = { workspace = true }
+futures = { workspace = true }
 linked_hash_set = { workspace = true }
 hickory-proto = { workspace = true }
 hickory-resolver = { workspace = true }
